Revert "Undo renames and moves"
