+++ web/status-history/src/main/java/net/twisterrob/blt/gapp/infrastructure/HttpStatusInteractor.java
@@ -1,4 +1,4 @@
-package net.twisterrob.blt.gapp;
+package net.twisterrob.blt.gapp.infrastructure;
 
 import java.io.IOException;
 import java.io.InputStream;
