+++ web/status-history/src/main/resources/views/index.hbs
@@ -12,9 +12,9 @@
 
 	<h2>Available apps</h2>
 	<ul>
-		<li><a href="refresh?feed=TubeDepartureBoardsLineStatus">FeedCronServlet (/refresh?feed=)</a></li>
-		<li><a href="LineStatusHistory?current=false&amp;errors=false">LineStatusHistoryServlet, just history, without errors</a></li>
-		<li><a href="LineStatusHistory?current=true&amp;errors=true">LineStatusHistoryServlet, with errors and current status</a></li>
+		<li><a href="refresh?feed=TubeDepartureBoardsLineStatus">RefreshFeedController (/refresh?feed=)</a></li>
+		<li><a href="LineStatusHistory?current=false&amp;errors=false">LineStatusHistoryController, just history, without errors</a></li>
+		<li><a href="LineStatusHistory?current=true&amp;errors=true">LineStatusHistoryController, with errors and current status</a></li>
 	</ul>
 
 	<h2>Version info</h2>
